first tests passed

+++ src/gamblegamble/IRandomValueGenerator.java
@@ -5,7 +5,7 @@ public interface IRandomValueGenerator {
 	
 
 	public double getRandomNum();
-	public double getRandomNumInRange(int min, int max);
+	public int getRandomNumInRange(int min, int max);
 	public boolean getHitProbability(double p);
